refactor: change namespacing to allow osu-localisation-analyzer to process all strings
