+++ osu.Game/Screens/Play/HUD/ArgonSongProgress.cs
@@ -8,7 +8,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Configuration;
 using osu.Game.Graphics;
-using osu.Game.Localisation.SkinEditorComponents;
+using osu.Game.Localisation;
 using osu.Game.Rulesets.Objects;
 
 namespace osu.Game.Screens.Play.HUD
