+++ osu.Game/Screens/Play/HUD/JudgementCounter/JudgementCounterDisplay.cs
@@ -8,7 +8,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Localisation;
 using osu.Game.Configuration;
-using osu.Game.Localisation.SkinEditorComponents;
+using osu.Game.Localisation;
 using osu.Game.Rulesets.Scoring;
 using osu.Game.Skinning;
 using osuTK;
