+++ osu.Game/Skinning/Components/TextElement.cs
@@ -8,7 +8,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Game.Configuration;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
-using osu.Game.Localisation.SkinEditorComponents;
+using osu.Game.Localisation;
 
 namespace osu.Game.Skinning.Components
 {
